# apps/api/docs/swagger.yaml
basePath: /
definitions:
  handlers.Criterias:
    properties:
      activity_types:
        items:
          type: string
        type: array
      daytime:
        items:
          type: string
        type: array
      departments:
        items:
          type: string
        type: array
      employee_ids:
        items:
          type: integer
        type: array
      goals:
        items:
          type: string
        type: array
      orgs:
        items:
          type: string
        type: array
      period:
        items:
          type: string
        type: array
    type: object
...
  model.Principal:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
    type: object
...
host: 188.72.77.222:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Реализованный в рамках хакатона сервер для корпоративной системы благотворительности
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: АПИ сервера для корпоративной системы благотворительности
  version: "1.0"
paths:
...
  /admin:
    post:
      consumes:
      - application/json
      description: Get one item
      parameters:
      - description: Principal
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Principal'
      produces:
      - application/json
      responses:
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Получение статистики
      tags:
      - stats
  /admin/login:
    post:
      consumes:
      - application/json
      description: Get one item
      parameters:
      - description: Pricipal
        in: body
        name: principal
        required: true
        schema:
          $ref: '#/definitions/model.Principal'
      produces:
      - application/json
      responses:
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Получение статистики
      tags:
      - stats
...
  /stats:
    get:
      consumes:
      - application/json
      description: Get one item
      parameters:
      - description: Principal
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handlers.Criterias'
      produces:
      - application/json
      responses:
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Получение статистики
      tags:
      - stats
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"